% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateInteractions.R
\name{CalculateInteractions}
\alias{CalculateInteractions}
\title{Raw Interaction Score Calculation}
\usage{
CalculateInteractions(
  counts_fn,
  metadata_fn,
  LR_pairs,
  pathway_components,
  perm_times,
  knowledgebase_version,
  delimitor,
  seuratObject
)
}
\arguments{
\item{counts_fn}{The file path of the gene/cell matrix containing all expression counts}

\item{metadata_fn}{The file path of the accompanying metadata file for counts}

\item{LR_pairs}{The ligand/receptor annotation object chosen based on the knowledgebase version}

\item{pathway_components}{The pathway core components object chosen based on the knowledgebase version}

\item{perm_times}{The number of times to shuffle the clusters and calculate ligand/receptor averages for pvalue calculations}

\item{knowledgebase_version}{The chosen knowledgebase version}

\item{delimitor}{The separator for the counts file if it is a .txt file. Default is tab.}

\item{seuratObject}{Counts/Metadata stored in a Seurat Object.}
}
\description{
Reads in user inputted counts/metadata file pairing and calculates
the raw interaction score for all cell-cell combinations.
}
\keyword{internal}
