% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateVisualizationsSingle.R
\name{ChordDiagramMulti}
\alias{ChordDiagramMulti}
\title{Chord diagram generation for Multi Analysis data}
\usage{
ChordDiagramMulti(output_dir = NULL, results, cellTypes = NULL)
}
\arguments{
\item{output_dir}{Directory to save generated heatmap(s) to. If unspecified, visualizations will not be saved and instead will simply be printed out.}

\item{results}{The file path to the results Excel file generated by RunMultiAnalysis()}

\item{cellTypes}{The celltypes of interest. Leave this unspecified if you want to explore all cell types.}
}
\description{
Chord diagram generation for Multi Analysis data
}
