% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateVisualizationsSingle.R
\name{CirclePlotMulti}
\alias{CirclePlotMulti}
\title{Circle plot generation for Multi Analysis data}
\usage{
CirclePlotMulti(
  output_dir = NULL,
  results,
  pathwayObj,
  pathways = NULL,
  cellTypes = NULL
)
}
\arguments{
\item{output_dir}{Directory to save generated visualizations to. If unspecified, visualizations will not be saved and instead will simply be printed out.}

\item{results}{The file path to the results Excel file generated by RunMultiAnalysis()}

\item{pathwayObj}{The core component file used in the analysis}

\item{pathways}{The pathways of interest. Leave this unspecified if you want to generate heatmaps for all pathways in pathwayObj.}

\item{cellTypes}{The celltypes of interest. Leave this unspecified if you want to explore all cell types.}
}
\description{
Circle plot generation for Multi Analysis data
}
