% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateVisualizations.R
\name{doMultiVisualization}
\alias{doMultiVisualization}
\title{Multi Analysis Visualization}
\usage{
doMultiVisualization(DEG, pathwayObj, base_output_dir)
}
\arguments{
\item{pathwayObj}{The core component file as determined by the user inputted knowledgebase version}

\item{base_output_dir}{The directory for FlyPhone to send all result files to. Defaults to the current working directory}

\item{DEG_fn}{The file path for the differentially expressed genes file}
}
\description{
Generates the following visualizations:
\itemize{
\item Log2FC Heatmap
\item Log2FC Circle Plot
\item Interaction score Chord Diagram
\item Interaction Strength Graph
}
}
\keyword{internal}
