% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateVisualizationsSingle.R
\name{DotPlotSingle}
\alias{DotPlotSingle}
\title{Dot plot generation for Single Analysis data}
\usage{
DotPlotSingle(output_dir = NULL, seuratObj, pathwayObj, pathways = NULL)
}
\arguments{
\item{output_dir}{Directory to save generated heatmap(s) to. If unspecified, visualizations will not be saved and instead will simply be printed out.}

\item{seuratObj}{A Seurat Object where counts = raw count data, and meta.data = metadata}

\item{pathwayObj}{The core component file used in the analysis}

\item{pathways}{The pathways of interest. Leave this unspecified if you want to generate heatmaps for all pathways in pathwayObj.}
}
\description{
Dot plot generation for Single Analysis data
}
