% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateVisualizations.R
\name{GenerateVisualizations}
\alias{GenerateVisualizations}
\title{Visualization Pipeline}
\usage{
GenerateVisualizations(
  counts_fn,
  metadata_fn,
  DEG,
  doMultivis,
  pathwayObj,
  delimitor,
  seuratObject,
  base_output_dir
)
}
\arguments{
\item{counts_fn}{The file path of the gene/cell matrix containing all expression counts}

\item{metadata_fn}{The file path of the accompanying metadata file for counts}

\item{DEG}{The differentially expressed gene object created in RunFlyPhone()}

\item{doMultivis}{Whether or not to generate comparison visualizations}

\item{pathwayObj}{The core component file as determined by the user inputted knowledgebase version}

\item{delimitor}{The separator for the counts file if it is a .txt file}

\item{seuratObject}{The file path to the Seurat Object formatted data}

\item{base_output_dir}{The directory for FlyPhone to send all result files to. Defaults to the current working directory}
}
\description{
Uses the inputted data and calculated results to create visualizations.
}
\keyword{internal}
