% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateVisualizationsSingle.R
\name{HeatmapMulti}
\alias{HeatmapMulti}
\title{Heatmap generation for Multi Analysis data}
\usage{
HeatmapMulti(output_dir = NULL, DEG, pathwayObj, pathways = NULL)
}
\arguments{
\item{output_dir}{Directory to save generated heatmap(s) to. If unspecified, visualizations will not be saved and instead will simply be printed out.}

\item{DEG}{The file path for the differentially expressed genes file}

\item{pathwayObj}{The core component file used in the analysis}

\item{pathways}{The pathways of interest. Leave this unspecified if you want to generate heatmaps for all pathways in pathwayObj.}
}
\description{
Heatmap generation for Multi Analysis data
}
