% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateVisualizationsSingle.R
\name{InteractionStrengthMulti}
\alias{InteractionStrengthMulti}
\title{Interaction Strength scatterplot generation for Multi Analysis data}
\usage{
InteractionStrengthMulti(output_dir = NULL, results)
}
\arguments{
\item{output_dir}{Directory to save generated heatmap(s) to. If unspecified, visualizations will not be saved and instead will simply be printed out.}

\item{results}{The file path to the results Excel file generated by RunMultiAnalysis()}
}
\description{
Interaction Strength scatterplot generation for Multi Analysis data
}
