% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateVisualizationsSingle.R
\name{InteractionStrengthSingle}
\alias{InteractionStrengthSingle}
\title{Interaction Strength scatterplot generation for Single Analysis data}
\usage{
InteractionStrengthSingle(output_dir = NULL, resultFileList)
}
\arguments{
\item{output_dir}{Directory to save generated visualizations to. If unspecified, visualizations will not be saved and instead will simply be printed out.}

\item{resultFileList}{The file path to the results file (long or long_filtered (recommended)) generated by RunSingleAnalysis(). If multiple files are inputted, InteractionStrengthSingle() will scale all axes to be the same.}
}
\description{
Interaction Strength scatterplot generation for Single Analysis data
}
