% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathway_components_v1}
\alias{pathway_components_v1}
\title{Pathway Core Components Version 1 - 2021}
\format{
\subsection{\verb{Pathway Components V1}}{

A data frame with 303 rows and 5 columns:
}
}
\source{
Manually curated
}
\usage{
pathway_components_v1
}
\description{
Dataset of core components information for each signaling pathway.
}
\keyword{datasets}
