% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathway_components_v2}
\alias{pathway_components_v2}
\title{Pathway Core Components Version 2.5 - 2025}
\format{
\subsection{\verb{Pathway Components V2}}{

A data frame with 458 rows and 6 columns:
}
}
\source{
Manually curated
}
\usage{
pathway_components_v2
}
\description{
Dataset of core components information for each signaling pathway.
}
\keyword{datasets}
