% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayELVisualizations.R
\name{PathwayVisualizations}
\alias{PathwayVisualizations}
\title{Pathway Expression Level Visualization Generator}
\usage{
PathwayVisualizations(output_dir = NULL, fileList)
}
\arguments{
\item{output_dir}{Directory to save visualizations to. If unspecified, visualizations will not be saved and instead will simply be printed out.}

\item{fileList}{A list of all the "pathway-expression-levels" CSV files generated by either RunSingleAnalysis() or RunMultiAnalysis().}
}
\description{
Generates visualizations for Pathway expression levels in each given sample
}
