% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{V1}
\alias{V1}
\title{FlyPhone Ligand-Receptor Pairs Version 1 - 2021}
\format{
\subsection{\verb{Version 1}}{

A data frame with 196 rows and 11 columns:
}
}
\source{
Manually curated
}
\usage{
V1
}
\description{
Dataset of all ligand-receptor pairs for the major fly signaling pathways.
}
\keyword{datasets}
