% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{V2A}
\alias{V2A}
\title{FlyPhone Ligand-Receptor Pairs Version 2.3 - 2025}
\format{
\subsection{\verb{Version 2 All}}{

A data frame with 1,853 rows and 22 columns:
}
}
\source{
Manually curated
}
\usage{
V2A
}
\description{
Dataset of all ligand-receptor pairs for the major fly signaling pathways.
}
\keyword{datasets}
