% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{V2H}
\alias{V2H}
\title{FlyPhone Ligand-Receptor Pairs Version 2.3 - 2025}
\format{
\subsection{\verb{Version 2 High}}{

A data frame with 259 rows and 22 columns:
}
}
\source{
Manually curated
}
\usage{
V2H
}
\description{
Dataset of ligand-receptor pairs for the major fly signaling pathways that have a rank of "High".
}
\keyword{datasets}
