% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{V2HM}
\alias{V2HM}
\title{FlyPhone Ligand-Receptor Pairs Version 2.3 - 2025}
\format{
\subsection{\verb{Version 2 High/Moderate}}{

A data frame with 906 rows and 22 columns:
}
}
\source{
Manually curated
}
\usage{
V2HM
}
\description{
Dataset of all ligand-receptor pairs for the major fly signaling pathways.
}
\keyword{datasets}
