% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunFlyPhone.R
\name{RunFlyPhone}
\alias{RunFlyPhone}
\title{Run FlyPhone pipeline}
\usage{
RunFlyPhone(
  counts_fn = NULL,
  metadata_fn = NULL,
  DEG = NULL,
  pct_filter = 0.1,
  knowledgebase_version,
  perm_times = 1000,
  delimitor = "\\t",
  seuratObject = NULL,
  control_name = NULL,
  mutant_name = NULL
)
}
\arguments{
\item{counts_fn}{The filepath to the counts matrix.}

\item{metadata_fn}{The filepath to the accompanying metadata file.}

\item{DEG}{The filepath to the DEG file, if provided.}

\item{pct_filter}{Percent threshold to filter expression of genes by. Default value of 0.1.}

\item{knowledgebase_version}{Version of knowledgebase to use. Valid values are: "Version 1", "Version 2 All", "Version 2 High", and "Version 2 High/Moderate".}

\item{perm_times}{The number of times to shuffle cluster alignments and calculate ligand/receptor averages for p-value calculation in raw interaction scoring. Default value of 1000.}

\item{delimitor}{The separator for the counts file if it is a .txt file. Default is tab.}

\item{control_name}{The name of the control sample, if applicable.}

\item{mutant_name}{The name of the mutant sample, if applicable.}
}
\description{
Wrapper function for the FlyPhone pipeline.
}
\examples{
# Afroditi data (GSE218641)
RunFlyPhone("GSE218641_filtered_feature_bc_matrix.txt", "GSE218641_GEO_metadata_Afroditi_Petsakou", knowledgebase_version = "Version 2 All")
RunFlyPhone(counts_fn = "GSE218641_filtered_feature_bc_matrix.txt", metadata_fn = "GSE218641_GEO_metadata_Afroditi_Petsakou", DEG = "Recovery_vs_Homeostasis_DEG.csv", pct_filter = 0.05, knowledgebase_version = "Version 2 All", perm_times = 100, delimitor = " ")

}
